requirements.txt
flask
tensorflow-gpu
tensorboard
pandas
opencv-python
matplotlib